fix: satisfy goose2 check gates
